[iOS] refactor : Response 프로토콜 기능을 새 파일로 이동
Associate with #26
